Created and added functionality for the new crate drop system using prefabs and spawners.

+++ Assets/Scripts/custom/BSpawner.js
@@ -1,23 +1,52 @@
 ﻿#pragma strict
 
+/*
+	Class Variables:
+	- XML File that will contain the beacon coordinates
+	- List of Vector3's that represent the beacon locations
+	- Boolean deciding whether or not to remove the beacons from the world
+	- Prefab object that will be cloned to create multiple beacons in the actual scene
+*/
 private var asset : String;
 private var list : List.<Vector3> = new List.<Vector3>();
 private var destroyBeacons : boolean = false;
 public var beaconPrefab : GameObject;
 
+/*
+	Start:
+	- Read in the initial trial of beacon data
+	- Spawn the beacons in the world
+*/
 function Start () {
 	readXMLData();
 	
 	instantiateBeacons();
 }
 
+/*
+	Update (Every frame):
+	- Check to see if the 'B' button is pressed
+	- If it is, toggle the boolean and either destroy or create the beacons
+*/
 function Update () {
 	if(Input.GetKeyDown(KeyCode.B)) {
 		destroyBeacons = !destroyBeacons;
-		destroyAllBeacons();
+
+		if(destroyBeacons) {
+			destroyAllBeacons();
+		}
+		else {
+			instantiateBeacons();
+		}
+
 	}
 }
 
+/*
+	Spawn Beacons:
+	- Traverses the list of beacon Vector3's and creates them from a prefab
+	- Sets the initial color to blue (will change later) and adds a tag for easy searching when destroying
+*/
 function instantiateBeacons() {
 	var newBeacon : GameObject;
 	for(position in list) {
@@ -27,17 +56,22 @@ function instantiateBeacons() {
 	}
 }
 
+/*
+	Destroy Beacons:
+	- Find all the game objects with tag "Beacon"
+	- Destroy all of those objects
+*/
 function destroyAllBeacons() {
-	if(destroyBeacons) {
-		for(beacon in GameObject.FindGameObjectsWithTag("Beacon")) {
-			Destroy(beacon);
-		}
-	}
-	else {
-		instantiateBeacons();
+	for(beacon in GameObject.FindGameObjectsWithTag("Beacon")) {
+		Destroy(beacon);
 	}
 }
 
+/*
+	XML Reader:
+	- Find the xml file named "TrialData.xml"
+	- Find all of the (x, y, z) values within the beacon tags
+*/
 function readXMLData() {
 	var tempV : Vector3;
